fix package of storage resolvers
