+++ app/src/main/java/org/stepic/droid/core/presenters/contracts/DownloadsView.kt
@@ -1,0 +1,4 @@
+package org.stepic.droid.core.presenters.contracts
+
+class DownloadsView {
+}
