--- app/src/main/java/org/stepic/droid/model/DownloadLink.kt
@@ -1,6 +1,0 @@
-package org.stepic.droid.model
-
-data class DownloadLink(
-        var architecture: String?,
-        var link: String?
-)
