+++ app/src/main/java/org/stepic/droid/persistence/downloads/resolvers/structure/UnitStructureResolver.kt
@@ -1,14 +1,13 @@
-package org.stepic.droid.persistence.downloads.adapters
+package org.stepic.droid.persistence.downloads.resolvers.structure
 
 import io.reactivex.Observable
 import org.stepic.droid.persistence.model.Structure
+import org.stepik.android.model.Unit
 
-interface StepStructureResolver {
+interface UnitStructureResolver: StructureResolver<Unit> {
     fun resolveStructure(
             courseId: Long,
             sectionId: Long,
-            unitId: Long,
-            lessonId: Long,
-            vararg stepIds: Long
+            vararg unitIds: Long
     ): Observable<Structure>
 }
